pyproject.toml
[tool.poetry]
name = "e-wallet"
version = "0.1.0"
description = ""
authors = ["bulatsabirow <bulatsabirow@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pathlib = "^1.0.1"
attrs = "^23.2.0"
arrow = "^1.3.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
